+++ deps/v8/src/debug/debug-property-iterator.h
@@ -6,9 +6,9 @@
 #define V8_DEBUG_DEBUG_PROPERTY_ITERATOR_H_
 
 #include "src/debug/debug-interface.h"
-#include "src/execution/isolate.h"
-#include "src/handles/handles.h"
-#include "src/objects/prototype.h"
+#include "src/handles.h"
+#include "src/isolate.h"
+#include "src/prototype.h"
 
 #include "include/v8.h"
 
@@ -46,9 +46,9 @@ class DebugPropertyIterator final : public debug::PropertyIterator {
   enum Stage { kExoticIndices = 0, kEnumerableStrings = 1, kAllProperties = 2 };
   Stage stage_ = kExoticIndices;
 
-  size_t current_key_index_ = 0;
+  uint32_t current_key_index_ = 0;
   Handle<FixedArray> keys_;
-  size_t exotic_length_ = 0;
+  uint32_t exotic_length_ = 0;
 
   bool calculated_native_accessor_flags_ = false;
   int native_accessor_flags_ = 0;
